feat: added network for solving 5 card hands. 5card hands are generated and shuffled to maintain randomness. Training on 5card works now as well

+++ poker/hand_recognition/data_utils.py
@@ -52,4 +52,4 @@ def unpack_nparrays(shape,batch,data):
             j += 1
         i += 1
     print('Numpy data uniques and counts ',np.lib.arraysetops.unique(Y,return_counts=True))
-    return torch.tensor(X),torch.tensor(Y).long()
+    return torch.tensor(X).float(),torch.tensor(Y).long()
